fix(RVCV-73): domain에 맞도록 api의 파라미터 수정(long -> Integer)

+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/application/usecase/ReviewUseCase.java
@@ -19,6 +19,6 @@ public interface ReviewUseCase {
 
 	void updateReview(long reviewId, UpdateReviewRequest updateReviewRequest);
 
-	PageResponse<GetAwaitReviewResponse> getAwaitReviewsByShopAdmin(long shopAdminId, PageableRequest pageableRequest);
-	
+	PageResponse<GetAwaitReviewResponse> getAwaitReviewsByShopAdmin(Integer shopAdminId, PageableRequest pageableRequest);
+
 }
